fix(objects): adopt size_t policy, RAII internals, remove unreachable code (-441 warnings via root-cause type policy, not suppression)
dbp_plugin_objects (+ dboformat dep) /W4 families eliminated: C4701 35->0 (init at declaration), C4456/C4458 35->0 (rename/remove shadowing), C4189 83->0 (dead locals deleted after side-effect review), C4244/C4267 159->0 (size_t receivers, DWORD only at binary/ABI boundaries), C4702 129->0 (unreachable code removed). Build green, dbp_tests 886/886.

+++ Dark Basic Public Shared/Dark Basic Pro SDK/Shared/Objects/Compiler/ProcessHSR.cpp
@@ -71,7 +71,7 @@ HRESULT cProcessHSR::Process ( void )
 	cBounds3    BoundsA,
 				BoundsB;
     cBSPTree**  pBSPTrees  = NULL;
-    int			iMeshCount = m_vpMeshList.size ( );
+    int			iMeshCount = static_cast<int>( m_vpMeshList.size ( ) );
     
 	memset ( &Options, 0, sizeof ( sBSP ) );
